Flexible Date Added
A custom flexible date type is added with id 10

So, Form Fields of type 10 means it is of the new flexible date type

Flexible dates allow user to enter just only year, or month and year.
Some parts of the date could be ignored.

A simple lists for day,month,year are shown up to help users but their
vales does’t saved on the DB. A hidden field collects their value and
saved on DB.