Date formatting issues
This version has a lot of changes related to date
1- submit date field in the submit form is disabled when changing and
the date picker is restricted to the past (until today). Users can’t
submit a report in date in the future

2- A custom date component is added where the user can put only the
month or year. it is not necessary to be a valid date format dd/mm/yyyy
(see themes/mrgTheme/views/reports/submit_custom_forms.php)

3- There is an error on submitting a report in the date. It goes back
to 1970 because the date reformatting is not correct and it set the
default date on the database (see application/helpers/reports.php)