Site TagLine and Massage in different languages
The problem is when language changed, the tagline and site message
doesn’t change to the required language. It is a database field where
you can change it in the settings. It is only in one language. It is
not allowed to add a different translation to the same message.
The solution is to write it in a certain format as follows

[(language) message]

example:
[(en_GB) A secure platform for reporting violations of the rights of
civilians in Iraq]
[(ar) منصة آمنة للإبلاغ عن انتهاكات حقوق المدنيين في العراق]