Enforce Address Geocoding on report Submit
To enforce users to locate the incident location on the map, The
incident location field is disabled until making search and if there is
more than one location (ambiguous address), then user may select the
correct one from the dropdown list under the search box.
Therefore, The searched location is copied to the incident location and
incident location is enabled.

We used our google API key to use google geocoding service.

Hint, The search is limited to Iraq , you can find that in the search
url.