Report Edit Map problem and governorates removal
A problem when report edit appears as the map doesn’t appear. It is
solved in this version (It is a problem in submit_edit_js.php)
Also, after making the new search, there is no need to the
Governorates, and cities dropdown list. So, they are removed in this
version.

Finally, the new search with geocoding and disambiguation dropdown list
are added to the edit report page.